+++ Src/Library/Endpoint/Endpoint.Setup.cs
@@ -3,7 +3,6 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System.Reflection;
 using System.Text.Json.Serialization;
-using static FastEndpoints.Config;
 
 namespace FastEndpoints;
 
@@ -69,8 +68,7 @@ public abstract partial class Endpoint<TRequest, TResponse> : BaseEndpoint where
     /// <summary>
     /// specify one or more route patterns this endpoint should be listening for
     /// </summary>
-    protected void Routes(params string[] patterns)
-        => Definition.Routes = patterns;
+    protected void Routes(params string[] patterns) => Definition.Routes = patterns;
 
     /// <summary>
     /// specify one or more http method verbs this endpoint should be accepting requests for
@@ -118,42 +116,32 @@ public abstract partial class Endpoint<TRequest, TResponse> : BaseEndpoint where
     /// disable auto validation failure responses (400 bad request with error details) for this endpoint.
     /// <para>HINT: this only applies to request dto validation.</para>
     /// </summary>
-    protected void DontThrowIfValidationFails()
-        => Definition.ThrowIfValidationFails = false;
+    protected void DontThrowIfValidationFails() => Definition.DontThrowIfValidationFails();
 
     /// <summary>
     /// use this only if you have your own exception catching middleware.
     /// if this method is called in config, an automatic error response will not be sent to the client by the library.
     /// all exceptions will be thrown and it would be your exeception catching middleware to handle them.
     /// </summary>
-    protected void DontCatchExceptions()
-        => Definition.DontCatchExceptions = true;
+    protected void DontCatchExceptions() => Definition.DontCatchExceptions();
 
     /// <summary>
     /// configure custom model binding for this endpoint by supplying an IRequestBinder implementation.
     /// by calling this method, you're completely bypassing the built-in model binding and taking things into your own hands for this endpoint.
     /// </summary>
     /// <param name="binder">custom model binder implementation to use for this endpoint</param>
-    protected void RequestBinder(IRequestBinder<TRequest> binder)
-        => Definition.RequestBinder = binder;
+    protected void RequestBinder(IRequestBinder<TRequest> binder) => Definition.RequestBinder = binder;
 
     /// <summary>
     /// if swagger auto tagging based on path segment is enabled, calling this method will prevent a tag from being added to this endpoint.
     /// </summary>
-    protected void DontAutoTag()
-        => Definition.DontAutoTag = true;
+    protected void DontAutoTag() => Definition.DontAutoTag();
 
     /// <summary>
     /// allow unauthenticated requests to this endpoint. optionally specify a set of verbs to allow unauthenticated access with.
     /// i.e. if the endpoint is listening to POST, PUT &amp; PATCH and you specify AllowAnonymous(Http.POST), then only PUT &amp; PATCH will require authentication.
     /// </summary>
-    protected void AllowAnonymous(params Http[] verbs)
-    {
-        Definition.AnonymousVerbs =
-            verbs.Length > 0
-            ? verbs.Select(v => v.ToString()).ToArray()
-            : Enum.GetNames(Types.Http);
-    }
+    protected void AllowAnonymous(params Http[] verbs) => Definition.AllowAnonymous(verbs);
 
     /// <summary>
     /// enable file uploads with multipart/form-data content type
@@ -162,92 +150,66 @@ public abstract partial class Endpoint<TRequest, TResponse> : BaseEndpoint where
     /// set 'true' to disable auto binding of form data which enables uploading and reading of large files without buffering to memory/disk.
     /// you can access the multipart sections for reading via the FormFileSectionsAsync() method.
     /// </param>
-    protected void AllowFileUploads(bool dontAutoBindFormData = false)
-    {
-        Definition.AllowFormData = true;
-        Definition.DontBindFormData = dontAutoBindFormData;
-    }
+    protected void AllowFileUploads(bool dontAutoBindFormData = false) => Definition.AllowFileUploads(dontAutoBindFormData);
 
     /// <summary>
     /// enable multipart/form-data submissions
     /// </summary>
-    protected void AllowFormData()
-        => Definition.AllowFormData = true;
+    protected void AllowFormData() => Definition.AllowFormData();
 
     /// <summary>
     /// specify one or more authorization policy names you have added to the middleware pipeline during app startup/ service configuration that should be applied to this endpoint.
     /// </summary>
     /// <param name="policyNames">one or more policy names (must have been added to the pipeline on startup)</param>
-    protected void Policies(params string[] policyNames)
-        => Definition.PreBuiltUserPolicies = policyNames;
+    protected void Policies(params string[] policyNames) => Definition.Policies(policyNames);
 
     /// <summary>
     /// allows access if the claims principal has ANY of the given roles
     /// </summary>
     /// <param name="rolesNames">one or more roles that has access</param>
-    protected void Roles(params string[] rolesNames)
-        => Definition.Roles = rolesNames;
+    protected void Roles(params string[] rolesNames) => Definition.Roles(rolesNames);
 
     /// <summary>
     /// allows access if the claims principal has ANY of the given permissions
     /// </summary>
     /// <param name="permissions">the permissions</param>
-    protected void Permissions(params string[] permissions)
-    {
-        Definition.AllowAnyPermission = true;
-        Definition.Permissions = permissions;
-    }
+    protected void Permissions(params string[] permissions) => Definition.Permissions(permissions);
 
     /// <summary>
     /// allows access if the claims principal has ALL of the given permissions
     /// </summary>
     /// <param name="permissions">the permissions</param>
-    protected void PermissionsAll(params string[] permissions)
-    {
-        Definition.AllowAnyPermission = false;
-        Definition.Permissions = permissions;
-    }
+    protected void PermissionsAll(params string[] permissions) => Definition.PermissionsAll(permissions);
 
     /// <summary>
     /// allows access if the claims principal has ANY of the given claim types
     /// </summary>
     /// <param name="claimTypes">the claim types</param>
-    protected void Claims(params string[] claimTypes)
-    {
-        Definition.AllowAnyClaim = true;
-        Definition.ClaimTypes = claimTypes;
-    }
+    protected void Claims(params string[] claimTypes) => Definition.Claims(claimTypes);
 
     /// <summary>
     /// allows access if the claims principal has ALL of the given claim types
     /// </summary>
     /// <param name="claimTypes">the claim types</param>
-    protected void ClaimsAll(params string[] claimTypes)
-    {
-        Definition.AllowAnyClaim = false;
-        Definition.ClaimTypes = claimTypes;
-    }
+    protected void ClaimsAll(params string[] claimTypes) => Definition.ClaimsAll(claimTypes);
 
     /// <summary>
     /// specify which authentication schemes to use for authenticating requests to this endpoint
     /// </summary>
     /// <param name="authSchemeNames">the authentication scheme names</param>
-    protected void AuthSchemes(params string[] authSchemeNames)
-        => Definition.AuthSchemes = authSchemeNames;
+    protected void AuthSchemes(params string[] authSchemeNames) => Definition.AuthSchemes(authSchemeNames);
 
     /// <summary>
     /// configure a collection of pre-processors to be executed before the main handler function is called. processors are executed in the order they are defined here.
     /// </summary>
     /// <param name="preProcessors">the pre processors to be executed</param>
-    protected void PreProcessors(params IPreProcessor<TRequest>[] preProcessors)
-        => Definition.PreProcessors = preProcessors;
+    protected void PreProcessors(params IPreProcessor<TRequest>[] preProcessors) => Definition.PreProcessors = preProcessors;
 
     /// <summary>
     /// configure a collection of post-processors to be executed after the main handler function is done. processors are executed in the order they are defined here.
     /// </summary>
     /// <param name="postProcessors">the post processors to be executed</param>
-    protected void PostProcessors(params IPostProcessor<TRequest, TResponse>[] postProcessors)
-        => Definition.PostProcessors = postProcessors;
+    protected void PostProcessors(params IPostProcessor<TRequest, TResponse>[] postProcessors) => Definition.PostProcessors = postProcessors;
 
     /// <summary>
     /// specify response caching settings for this endpoint
@@ -257,23 +219,13 @@ public abstract partial class Endpoint<TRequest, TResponse> : BaseEndpoint where
     /// <param name="noStore">specify whether the data should be stored or not</param>
     /// <param name="varyByHeader">the value for the Vary response header</param>
     /// <param name="varyByQueryKeys">the query keys to vary by</param>
-    protected void ResponseCache(int durationSeconds, ResponseCacheLocation location = ResponseCacheLocation.Any, bool noStore = false, string? varyByHeader = null, string[]? varyByQueryKeys = null)
-    {
-        Definition.ResponseCacheSettings = new()
-        {
-            Duration = durationSeconds,
-            Location = location,
-            NoStore = noStore,
-            VaryByHeader = varyByHeader,
-            VaryByQueryKeys = varyByQueryKeys
-        };
-    }
+    protected void ResponseCache(int durationSeconds, ResponseCacheLocation location = ResponseCacheLocation.Any, bool noStore = false, string? varyByHeader = null, string[]? varyByQueryKeys = null) => Definition.ResponseCache(durationSeconds, location, noStore, varyByHeader, varyByQueryKeys);
 
     /// <summary>
     /// set endpoint configurations options using an endpoint builder action ///
     /// </summary>
     /// <param name="builder">the builder for this endpoint</param>
-    protected void Options(Action<RouteHandlerBuilder> builder) => Definition.UserConfigAction = builder;
+    protected void Options(Action<RouteHandlerBuilder> builder) => Definition.Options(builder);
 
     /// <summary>
     /// describe openapi metadata for this endpoint. optionaly specify whether or not you want to clear the default Accepts/Produces metadata.
@@ -283,46 +235,32 @@ public abstract partial class Endpoint<TRequest, TResponse> : BaseEndpoint where
     /// </summary>
     /// <param name="builder">the route handler builder for this endpoint</param>
     /// <param name="clearDefaults">set to true if the defaults should be cleared</param>
-    protected void Description(Action<RouteHandlerBuilder> builder, bool clearDefaults = false)
-        => Definition.UserConfigAction = clearDefaults
-                                         ? ClearDefaultAcceptsProducesMetadata + builder
-                                         : builder;
+    protected void Description(Action<RouteHandlerBuilder> builder, bool clearDefaults = false) => Definition.Description(builder, clearDefaults);
 
     /// <summary>
     /// provide a summary/description for this endpoint to be used in swagger/ openapi
     /// </summary>
     /// <param name="endpointSummary">an action that sets values of an endpoint summary object</param>
-    protected void Summary(Action<EndpointSummary> endpointSummary)
-    {
-        Definition.Summary = new();
-        endpointSummary(Definition.Summary);
-    }
+    protected void Summary(Action<EndpointSummary> endpointSummary) => Definition.Summary(endpointSummary);
 
     /// <summary>
     /// provide a summary/description for this endpoint to be used in swagger/ openapi
     /// </summary>
     /// <param name="endpointSummary">an action that sets values of an endpoint summary object</param>
-    protected void Summary(Action<EndpointSummary<TRequest>> endpointSummary)
-    {
-        var summary = new EndpointSummary<TRequest>();
-        endpointSummary(summary);
-        Definition.Summary = summary;
-    }
+    protected void Summary(Action<EndpointSummary<TRequest>> endpointSummary) => Definition.Summary(endpointSummary);
 
     /// <summary>
     /// provide a summary/description for this endpoint to be used in swagger/ openapi
     /// </summary>
     /// <param name="endpointSummary">an endpoint summary instance</param>
-    protected void Summary(EndpointSummary endpointSummary)
-        => Definition.Summary = endpointSummary;
+    protected void Summary(EndpointSummary endpointSummary) => Definition.Summary(endpointSummary);
 
     /// <summary>
     /// specify one or more string tags for this endpoint so they can be used in the exclusion filter during registration.
     /// <para>HINT: these tags have nothing to do with swagger tags!</para>
     /// </summary>
     /// <param name="endpointTags">the tag values to associate with this endpoint</param>
-    protected void Tags(params string[] endpointTags)
-        => Definition.Tags = endpointTags;
+    protected void Tags(params string[] endpointTags) => Definition.Tags(endpointTags);
 
     /// <summary>
     /// specify the version of the endpoint if versioning is enabled
@@ -331,16 +269,8 @@ public abstract partial class Endpoint<TRequest, TResponse> : BaseEndpoint where
     /// <param name="deprecateAt">the version group number starting at which this endpoint should not be included in swagger</param>
     protected void Version(int version, int? deprecateAt = null)
     {
-        Definition.Version.Current = GetVersion(version);
+        Definition.Version.Current = version;
         Definition.Version.DeprecatedAt = deprecateAt ?? 0;
-
-        static int GetVersion(int epVer)
-        {
-            return
-                epVer is 0 && VersioningOpts is not null
-                ? VersioningOpts.DefaultVersion
-                : epVer;
-        }
     }
 
     /// <summary>
@@ -353,21 +283,18 @@ public abstract partial class Endpoint<TRequest, TResponse> : BaseEndpoint where
     /// header name can also be configured globally using <c>app.UseFastEndpoints(c=> c.ThrottleOptions...)</c>
     /// not specifying a header name will first look for 'X-Forwarded-For' header and if not present, will use `HttpContext.Connection.RemoteIpAddress`.
     /// </param>
-    protected void Throttle(int hitLimit, double durationSeconds, string? headerName = null)
-        => Definition.HitCounter = new(headerName, durationSeconds, hitLimit);
+    protected void Throttle(int hitLimit, double durationSeconds, string? headerName = null) => Definition.Throttle(hitLimit, durationSeconds, headerName);
 
     /// <summary>
     /// specify the json serializer context if code generation for request/response dtos is being used
     /// </summary>
     /// <typeparam name="TContext">the type of the json serializer context for this endpoint</typeparam>
-    protected void SerializerContext<TContext>(TContext serializerContext) where TContext : JsonSerializerContext
-        => Definition.SerializerContext = serializerContext;
+    protected void SerializerContext<TContext>(TContext serializerContext) where TContext : JsonSerializerContext => Definition.SerializerContext = serializerContext;
 
     /// <summary>
     /// register the validator for this endpoint as scoped instead of singleton. which will enable constructor injection at the cost of performance.
     /// </summary>
-    protected void ScopedValidator()
-        => Definition.ScopedValidator = true;
+    protected void ScopedValidator() => Definition.ScopedValidator();
 
     /// <summary>
     /// specify an override route prefix for this endpoint if a global route prefix is enabled.
@@ -375,6 +302,5 @@ public abstract partial class Endpoint<TRequest, TResponse> : BaseEndpoint where
     /// global prefix can be ignored by setting <c>string.Empty</c>
     /// </summary>
     /// <param name="routePrefix">route prefix value</param>
-    protected void RoutePrefixOverride(string routePrefix)
-        => Definition.RoutePrefixOverride = routePrefix;
+    protected void RoutePrefixOverride(string routePrefix) => Definition.RoutePrefixOverride(routePrefix);
 }